Extract test_spider_crawl.py.
